My suggestions for changes.

+++ src/main/java/com/jindero/banking/features/account/Chargeable.java
@@ -1,8 +1,10 @@
 package com.jindero.banking.features.account;
 
+import java.math.BigDecimal;
+
 public interface Chargeable {
 
-    double calculateFees();
+    BigDecimal calculateFees();
 
     void applyMonthlyFee();
 }
